Rename overlay2Example.tsx
