+++ packages/docs-app/src/examples/core-examples/index.ts
@@ -63,7 +63,7 @@ export * from "./numericInputBasicExample";
 export * from "./numericInputExtendedExample";
 export * from "./nonIdealStateExample";
 export * from "./overflowListExample";
-export { OverlayExample } from "./overlay2Example";
+export { OverlayExample } from "./overlayExample";
 export { PanelStack2Example } from "./panelStack2Example";
 export * from "./popoverDismissExample";
 export * from "./popoverExample";
